Rename #79 - Pages 폴더 구조 변경
