+++ src/pages/Chat/LocationSharePage/LocationSharePageStyled.tsx
@@ -1,9 +1,12 @@
 import styled from '@emotion/styled';
 
 export const LocationSharePageStyled = styled.div`
+  padding: 50px 20px 0px;
+  width: 100%;
+  height: 100vh;
   display: flex;
   flex-direction: column;
-  flex: 1;
+  overflow: hidden;
   .map-wrap {
     flex: 1;
     margin: 0px -20px;
