+++ src/pages/Product/NewProductPage/NewProductPageStyled.tsx
