+++ src/pages/Profile/BuyHistoryPage/BuyHistoryPageStyled.tsx
