+++ src/pages/Profile/LikeListPage/LikeListPage.tsx
@@ -1,7 +1,7 @@
 import { LikeListPageStyled } from './LikeListPageStyled';
 import Top from '@molecules/Top/Top';
 import MyProductBox from '@molecules/MyProductBox/MyProductBox';
-import { dummyProduct } from 'pages/ProductDetailPage/ProductDetailPage';
+import { dummyProduct } from 'pages/Product/ProductDetailPage/ProductDetailPage';
 
 const LikeListPage = () => {
   return (
