+++ src/pages/Profile/MannerPage/MannerPageStyled.tsx
